requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai-whisper==20231117
torch>=2.0.0
python-multipart
aiofiles
python-dotenv
jinja2
pydantic
numpy
ffmpeg-python
loguru
tqdm
